🚚 move meta content to the devtools module
