+++ packages/atom.io/src/internal/index.ts
@@ -2,7 +2,6 @@ export * from "./atom-internal"
 export * from "./families-internal"
 export * from "./get"
 export * from "./is-default"
-export * as META from "./meta"
 export * from "./operation"
 export * from "./selector-internal"
 export * from "./selector"
