--- packages/atom.io/src/internal/meta/meta-transactions.ts
@@ -1,1 +1,0 @@
-export {}
